01-app/requirements.txt
# for local development
python-dotenv

# for local development & cloud function
google-cloud-storage
google-cloud-firestore
google-cloud-aiplatform
PyPDF2